UPDATE: Specific routers file, reduce directories, removed dockerfile and docker compose
