--- .env.example
@@ -1,6 +1,0 @@
-DATABASE_URL=<SUA_URL_DE_CONEXÃO>
-SG_KEY=<SUA_API_KEY>
-SG_TEMPLATE_ID=<SEU_TEMPLATE_ID>
-TWILIO_PHONE=<SEU_TELEFONE>
-TWILIO_SID=<SEU_SID>
-TWILIO_TOKEN=<SEU_TOKEN>
