+++ .gitignore
@@ -21,3 +21,5 @@ vendor/*
 db/*
 
 .env
+
+fiscaliza.db
