--- docker-compose.yaml
@@ -1,30 +1,0 @@
-services:
-  db:
-    image: postgres:13
-    container_name: db
-    restart: always
-    environment:
-      POSTGRES_USER: root
-      POSTGRES_PASSWORD: SUA_SENHA
-      POSTGRES_DB: fiscaliza
-    ports:
-      - "5432:5432"
-    volumes:
-      - ./db:/var/lib/postgresql/data
-
-  app:
-    build:
-        context: .
-        dockerfile: Dockerfile
-    container_name: fiscaliza_app
-    ports:
-      - "8000:8000"
-    depends_on:
-      - db
-    environment:
-      - DATABASE_URL=postgres://root:<SUA_SENHA>@db:5432/fiscaliza?sslmode=disable
-
-    command: ["./main"]
-
-volumes:
-    db:
