--- internal/crypt/main.go
@@ -1,19 +1,0 @@
-package crypt
-
-import (
-	"fmt"
-	"golang.org/x/crypto/bcrypt"
-)
-
-func Password(password string) string {
-	hash, err := bcrypt.GenerateFromPassword([]byte(password), bcrypt.DefaultCost)
-	if err != nil {
-		fmt.Println(err)
-	}
-	return string(hash)
-}
-
-func ComparePassword(hashedPassword, password string) bool {
-	err := bcrypt.CompareHashAndPassword([]byte(hashedPassword), []byte(password))
-	return err == nil
-}
